+++ utilities/simpleTimer.hpp
@@ -29,7 +29,7 @@ public:
 
     double getSec() const
     {
-        double timePassed = elapsed().count();
+        double timePassed = static_cast<double>(elapsed().count());
         timePassed *= 1e-6;
         return timePassed;
     }
